# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Meemo API
  version: 2.0.0
  description: >
    Meemo is a personal note and knowledge management service.
    This specification describes the authoritative HTTP REST API for Meemo v2.
servers:
  - url: /
    description: Current host server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie generated after successful login.
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: Bad Request
        code:
          type: string
          example: invalid_request
          enum:
            - invalid_request
            - authentication_required
            - invalid_credentials
            - forbidden
            - not_found
            - conflict
            - payload_too_large
            - too_many_requests
            - internal_error
            - service_unavailable
            - request_failed
        message:
          type: string
          example: content must be a string
...
    AttachmentDescriptor:
      type: object
      required:
        - identifier
      properties:
        identifier:
          type: string
          description: Opaque storage UUID
          example: 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d
        fileName:
          type: string
          example: screenshot.png
        type:
          type: string
          enum: [image, unknown]
          example: image
        mime:
          type: string
          example: image/png
        size:
          type: integer
          example: 1048576
    ExternalContentItem:
      type: object
      required:
        - url
        - type
      properties:
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [image, unknown]
    NoteColor:
      type: string
      enum: [default, coral, peach, sand, mint, sage, fog, storm, dusk, blossom, clay, chalk]
      example: default
    Thing:
      type: object
      required:
        - _id
        - ownerId
        - content
        - richContent
        - createdAt
        - modifiedAt
        - tags
        - attachments
        - public
        - shared
        - archived
        - sticky
      properties:
        _id:
          type: string
          example: 64f8a12b3c4d5e6f7a8b9c0d
        ownerId:
          type: string
          example: 64f8a12b3c4d5e6f7a8b9c01
        content:
          type: string
          example: Hello #meeting with @bob
        richContent:
          type: string
          example: Hello <a href="/tags/meeting">#meeting</a> with @bob
        createdAt:
          type: integer
          description: Unix timestamp in milliseconds
          example: 1694000000000
        modifiedAt:
          type: integer
          description: Unix timestamp in milliseconds of the latest complete note update
          example: 1694000000000
        tags:
          type: array
          items:
            type: string
          example: [meeting]
        externalContent:
          type: array
          items:
            $ref: '#/components/schemas/ExternalContentItem'
          default: []
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentDescriptor'
        public:
          type: boolean
          example: false
        shared:
          type: boolean
          example: false
        archived:
          type: boolean
          example: false
        sticky:
          type: boolean
          example: false
        color:
          allOf:
            - $ref: '#/components/schemas/NoteColor'
          default: default
        score:
          type: number
          description: Relevance match score when search mode is text
          example: 1.25
    Tag:
      type: object
      required:
        - name
        - usage
      properties:
        name:
          type: string
          example: meeting
        usage:
          type: integer
          example: 5
...
    CreateThingRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          example: Quick note with #idea
        attachments:
          type: array
          items:
            oneOf:
              - type: string
                description: Attachment identifier or safe path segment
              - $ref: '#/components/schemas/AttachmentDescriptor'
          default: []
        color:
          allOf:
            - $ref: '#/components/schemas/NoteColor'
          default: default
    UpdateThingRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          example: Updated note with #idea
        attachments:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/AttachmentDescriptor'
          default: []
        public:
          type: boolean
          default: false
        shared:
          type: boolean
          default: false
        archived:
          type: boolean
          default: false
        sticky:
          type: boolean
          default: false
        color:
          allOf:
            - $ref: '#/components/schemas/NoteColor'
          example: coral
...
  parameters:
    ThingIdPath:
      name: id
      in: path
      required: true
      description: 24-character hexadecimal ObjectId of the Thing
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 64f8a12b3c4d5e6f7a8b9c0d
...
    FilterQuery:
      name: filter
      in: query
      required: false
      description: Search query string supporting text keywords, #tags, and [[WikiLinks]]
      schema:
        type: string
        maxLength: 1000
    SearchModeQuery:
      name: mode
      in: query
      required: false
      description: Search execution mode ('regex' for substring pattern matching, 'text' for MongoDB full-text index with stemming and scoring)
      schema:
        type: string
        enum: [regex, text]
        default: regex
    StickyQuery:
      name: sticky
      in: query
      required: false
      description: Filter by sticky flag
      schema:
        type: boolean
        default: false
    ArchivedQuery:
      name: archived
      in: query
      required: false
      description: Filter by archived flag
      schema:
        type: boolean
        default: false
    SkipQuery:
      name: skip
      in: query
      required: false
      description: Number of records to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitQuery:
      name: limit
      in: query
      required: false
      description: Maximum records to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    BadRequest:
      description: Validation or parameter format error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/things:
    get:
      tags: [things]
      summary: List user notes
      description: Returns paginated notes with optional text search, sticky, and archive filters.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/FilterQuery'
        - $ref: '#/components/parameters/SearchModeQuery'
        - $ref: '#/components/parameters/StickyQuery'
        - $ref: '#/components/parameters/ArchivedQuery'
        - $ref: '#/components/parameters/SkipQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: List of notes
          content:
            application/json:
              schema:
                type: object
                required: [things]
                properties:
                  things:
                    type: array
                    items:
                      $ref: '#/components/schemas/Thing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [things]
      summary: Create a note
      description: Creates a new note, extracting tags and processing markdown rich content.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThingRequest'
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                type: object
                required: [thing]
                properties:
                  thing:
                    $ref: '#/components/schemas/Thing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/things/{id}:
    get:
      tags: [things]
      summary: Get a note by ID
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ThingIdPath'
      responses:
        '200':
          description: Note details
          content:
            application/json:
              schema:
                type: object
                required: [thing]
                properties:
                  thing:
                    $ref: '#/components/schemas/Thing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [things]
      summary: Update a note
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ThingIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateThingRequest'
      responses:
        '201':
          description: Note updated successfully
          content:
            application/json:
              schema:
                type: object
                required: [thing]
                properties:
                  thing:
                    $ref: '#/components/schemas/Thing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [things]
      summary: Delete a note
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ThingIdPath'
      responses:
        '200':
          description: Note deleted successfully
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tags:
    get:
      tags: [things]
      summary: List user tags
      description: Returns distinct tags and their usage counts for the authenticated user.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: object
                required: [tags]
                properties:
                  tags:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...